feat(auth): Rename provider folder to auth
