+++ tsconfig.json
@@ -22,7 +22,7 @@
     "src/utils/firebase_object_factory.spec.ts",
     "src/utils/firebase_list_observable.spec.ts",
     "src/utils/firebase_object_observable.spec.ts",
-    "src/providers/auth.spec.ts",
+    "src/auth/auth.spec.ts",
     "typings/main.d.ts"
   ]
 }
